fix: refactor adapters, gsonutils, revert to standart response text
