+++ src/fun/ntony4u/kanban/server/BaseHttpHandler.java
@@ -1,34 +1,25 @@
 package fun.ntony4u.kanban.server;
 
 import com.google.gson.Gson;
-import com.google.gson.GsonBuilder;
 import com.sun.net.httpserver.HttpExchange;
 import com.sun.net.httpserver.HttpHandler;
 import fun.ntony4u.kanban.service.TaskManager;
-import fun.ntony4u.kanban.utils.DurationTypeAdapter;
-import fun.ntony4u.kanban.utils.LocalDateTimeTypeAdapter;
+import fun.ntony4u.kanban.utils.*;
 
 import java.io.IOException;
 import java.io.OutputStream;
+import java.net.HttpURLConnection;
 import java.nio.charset.StandardCharsets;
-import java.time.Duration;
-import java.time.LocalDateTime;
+import java.util.Map;
 
 public abstract class BaseHttpHandler implements HttpHandler {
     protected final TaskManager taskManager;
-    protected final Gson gson = new GsonBuilder()
-            .registerTypeAdapter(Duration.class, new DurationTypeAdapter())
-            .registerTypeAdapter(LocalDateTime.class, new LocalDateTimeTypeAdapter())
-            .create();
-
-    public static Gson getGson() {
-        return new GsonBuilder()
-                .registerTypeAdapter(Duration.class, new DurationTypeAdapter())
-                .registerTypeAdapter(LocalDateTime.class, new LocalDateTimeTypeAdapter())
-                .create();
-    }
+    protected final Gson gson = GsonUtils.getGson();
 
     protected BaseHttpHandler(TaskManager taskManager) {
+        if (taskManager == null) {
+            throw new IllegalArgumentException("TaskManager cannot be null");
+        }
         this.taskManager = taskManager;
     }
 
@@ -43,26 +34,33 @@ public abstract class BaseHttpHandler implements HttpHandler {
 
     protected void sendSuccess(HttpExchange exchange, Object responseObject) throws IOException {
         String response = gson.toJson(responseObject);
-        sendText(exchange, response, 200);
+        sendText(exchange, response, HttpURLConnection.HTTP_OK);
     }
 
     protected void sendCreated(HttpExchange exchange, Object responseObject) throws IOException {
         String response = gson.toJson(responseObject);
-        sendText(exchange, response, 201);
+        sendText(exchange, response, HttpURLConnection.HTTP_CREATED);
     }
 
     protected void sendNotFound(HttpExchange exchange) throws IOException {
-        sendText(exchange, "Объект не найден", 404);
+        String errorJson = gson.toJson(Map.of("error", "Object not found"));
+        sendText(exchange, errorJson, HttpURLConnection.HTTP_NOT_FOUND);
     }
 
     protected void sendHasInteractions(HttpExchange exchange) throws IOException {
-        sendText(exchange, "Задача пересекается по времени с существующими", 406);
+        String errorJson = gson.toJson(Map.of(
+                "error", "Task time overlaps with existing tasks",
+                "code", "time_overlap"
+        ));
+        sendText(exchange, errorJson, HttpURLConnection.HTTP_NOT_ACCEPTABLE);
     }
 
     protected void sendInternalError(HttpExchange exchange, Exception e) throws IOException {
-        e.printStackTrace();
-        String errorMsg = "Внутренняя ошибка сервера: " + e.getMessage();
-        sendText(exchange, errorMsg, 500);
+        String errorJson = gson.toJson(Map.of(
+                "error", "Internal server error",
+                "details", e.getMessage()
+        ));
+        sendText(exchange, errorJson, HttpURLConnection.HTTP_INTERNAL_ERROR);
     }
 
     protected String readRequestBody(HttpExchange exchange) throws IOException {
